Refactor OLEDSNGProcessor
