+++ ole-docstore/ole-docstore-webapp/src/main/java/org/kuali/ole/dsng/rest/processor/OleDsNgOverlayProcessor.java
@@ -10,7 +10,6 @@ import org.codehaus.jackson.type.TypeReference;
 import org.codehaus.jettison.json.JSONArray;
 import org.codehaus.jettison.json.JSONException;
 import org.codehaus.jettison.json.JSONObject;
-import org.eclipse.jetty.util.ajax.JSON;
 import org.kuali.ole.constants.OleNGConstants;
 import org.kuali.ole.docstore.common.constants.DocstoreConstants;
 import org.kuali.ole.docstore.common.document.EHoldings;
@@ -29,21 +28,8 @@ import org.kuali.ole.dsng.model.HoldingsRecordAndDataMapping;
 import org.kuali.ole.dsng.model.ItemRecordAndDataMapping;
 import org.kuali.ole.dsng.rest.Exchange;
 import org.kuali.ole.dsng.rest.handler.Handler;
-import org.kuali.ole.dsng.rest.handler.bib.CreateBibHandler;
-import org.kuali.ole.dsng.rest.handler.bib.DiscardBibHandler;
-import org.kuali.ole.dsng.rest.handler.bib.UpdateBibHandler;
-import org.kuali.ole.dsng.rest.handler.eholdings.*;
-import org.kuali.ole.dsng.rest.handler.holdings.CallNumberHandler;
-import org.kuali.ole.dsng.rest.handler.holdings.CallNumberPrefixHandler;
-import org.kuali.ole.dsng.rest.handler.holdings.CallNumberTypeHandler;
-import org.kuali.ole.dsng.rest.handler.holdings.CopyNumberHandler;
 import org.kuali.ole.dsng.rest.handler.holdings.*;
 import org.kuali.ole.dsng.rest.handler.items.*;
-import org.kuali.ole.dsng.rest.handler.items.DonorCodeHandler;
-import org.kuali.ole.dsng.rest.handler.items.DonorNoteHandler;
-import org.kuali.ole.dsng.rest.handler.items.DonorPublicDisplayHandler;
-import org.kuali.ole.dsng.rest.handler.items.StatisticalSearchCodeHandler;
-import org.kuali.ole.dsng.util.OleDsHelperUtil;
 import org.marc4j.marc.Record;
 import org.springframework.beans.factory.annotation.Autowired;
 
@@ -53,7 +39,7 @@ import java.util.*;
 /**
  * Created by SheikS on 12/8/2015.
  */
-public class OleDsNgOverlayProcessor extends OleDsHelperUtil implements DocstoreConstants {
+public class OleDsNgOverlayProcessor extends OleDsNgOverlayProcessorHelper implements DocstoreConstants {
     @Autowired
     BibDAO bibDAO;
 
@@ -63,185 +49,34 @@ public class OleDsNgOverlayProcessor extends OleDsHelperUtil implements Docstore
     @Autowired
     ItemDAO itemDAO;
 
-    private List<Handler> bibHandlers;
-    private List<Handler> holdingHandlers;
-    private List<Handler> eholdingHandlers;
-    private List<Handler> itemHandlers;
-    private List<HoldingsHandler> holdingMetaDataHandlers;
-    private List<HoldingsHandler> eholdingMetaDataHandlers;
-    private List<ItemHandler> itemMetaDataHandlers;
-
-    public List<Handler> getBibHandlers() {
-        if (null == bibHandlers) {
-            bibHandlers = new ArrayList<Handler>();
-            bibHandlers.add(new CreateBibHandler());
-            bibHandlers.add(new UpdateBibHandler());
-            bibHandlers.add(new DiscardBibHandler());
-        }
-        return bibHandlers;
-    }
-
-    public List<HoldingsHandler> getHoldingMetaDataHandlers() {
-        if (null == holdingMetaDataHandlers) {
-            holdingMetaDataHandlers = new ArrayList<HoldingsHandler>();
-            holdingMetaDataHandlers.add(new HoldingsLocationHandler());
-            holdingMetaDataHandlers.add(new CallNumberHandler());
-            holdingMetaDataHandlers.add(new CallNumberTypeHandler());
-            holdingMetaDataHandlers.add(new CallNumberPrefixHandler());
-            holdingMetaDataHandlers.add(new CopyNumberHandler());
-        }
-        return holdingMetaDataHandlers;
-    }
-
-    public List<ItemHandler> getItemMetaDataHandlers() {
-        if (null == itemMetaDataHandlers) {
-            itemMetaDataHandlers = new ArrayList<ItemHandler>();
-            itemMetaDataHandlers.add(new org.kuali.ole.dsng.rest.handler.items.CallNumberHandler());
-            itemMetaDataHandlers.add(new org.kuali.ole.dsng.rest.handler.items.CallNumberPrefixHandler());
-            itemMetaDataHandlers.add(new org.kuali.ole.dsng.rest.handler.items.CallNumberTypeHandler());
-            itemMetaDataHandlers.add(new ChronologyHandler());
-            itemMetaDataHandlers.add(new org.kuali.ole.dsng.rest.handler.items.CopyNumberHandler());
-            itemMetaDataHandlers.add(new DonorCodeHandler());
-            itemMetaDataHandlers.add(new DonorNoteHandler());
-            itemMetaDataHandlers.add(new DonorPublicDisplayHandler());
-            itemMetaDataHandlers.add(new EnumerationHandler());
-            itemMetaDataHandlers.add(new ItemHoldingLocationHandler());
-            itemMetaDataHandlers.add(new ItemBarcodeHandler());
-            itemMetaDataHandlers.add(new ItemStatusHandler());
-            itemMetaDataHandlers.add(new ItemTypeHandler());
-            itemMetaDataHandlers.add(new ItemLocationHandler());
-            itemMetaDataHandlers.add(new StatisticalSearchCodeHandler());
-            itemMetaDataHandlers.add(new VendorLineItemIdHandler());
-        }
-        return itemMetaDataHandlers;
-    }
-
-    public List<HoldingsHandler> getEholdingMetaDataHandlers() {
-
-        if (null == eholdingMetaDataHandlers) {
-            eholdingMetaDataHandlers = new ArrayList<HoldingsHandler>();
-            eholdingMetaDataHandlers.add(new HoldingsLocationHandler());
-            eholdingMetaDataHandlers.add(new CallNumberHandler());
-            eholdingMetaDataHandlers.add(new CallNumberTypeHandler());
-            eholdingMetaDataHandlers.add(new CallNumberPrefixHandler());
-            eholdingMetaDataHandlers.add(new CopyNumberHandler());
-            eholdingMetaDataHandlers.add(new AccessLocationHandler());
-            eholdingMetaDataHandlers.add(new AccessPasswordHandler());
-            eholdingMetaDataHandlers.add(new AccessStatusHandler());
-            eholdingMetaDataHandlers.add(new AccessUserNameHandler());
-            eholdingMetaDataHandlers.add(new AccessPasswordHandler());
-            eholdingMetaDataHandlers.add(new AdminUrlHandler());
-            eholdingMetaDataHandlers.add(new AdminUserNameHandler());
-            eholdingMetaDataHandlers.add(new AuthenticationTypeHandler());
-            eholdingMetaDataHandlers.add(new CancellationDecisionDateHandler());
-            eholdingMetaDataHandlers.add(new CancellationEffectiveDateHandler());
-            eholdingMetaDataHandlers.add(new CancellationReasonHandler());
-            eholdingMetaDataHandlers.add(new CoverageStartDateHandler());
-            eholdingMetaDataHandlers.add(new CoverageStartIssueHandler());
-            eholdingMetaDataHandlers.add(new CoverageStartVolumeHandler());
-            eholdingMetaDataHandlers.add(new CoverageEndDateHandler());
-            eholdingMetaDataHandlers.add(new CoverageEndIssueHandler());
-            eholdingMetaDataHandlers.add(new CoverageEndVolumeHandler());
-            eholdingMetaDataHandlers.add(new CurrentSubscriptionEndDateHandler());
-            eholdingMetaDataHandlers.add(new CurrentSubscriptionStartDateHandler());
-            eholdingMetaDataHandlers.add(new org.kuali.ole.dsng.rest.handler.eholdings.DonorCodeHandler());
-            eholdingMetaDataHandlers.add(new org.kuali.ole.dsng.rest.handler.eholdings.DonorNoteHandler());
-            eholdingMetaDataHandlers.add(new org.kuali.ole.dsng.rest.handler.eholdings.DonorPublicDisplayHandler());
-            eholdingMetaDataHandlers.add(new EResourceIdHandler());
-            eholdingMetaDataHandlers.add(new NoOfSumultaneousUserHander());
-            eholdingMetaDataHandlers.add(new InitialSubscriptionEndDateHandler());
-            eholdingMetaDataHandlers.add(new PerpetualAccessStartDateHandler());
-            eholdingMetaDataHandlers.add(new PerpetualAccessStartIssueHandler());
-            eholdingMetaDataHandlers.add(new PerpetualAccessStartVolumeHandler());
-            eholdingMetaDataHandlers.add(new PerpetualAccessEndDateHandler());
-            eholdingMetaDataHandlers.add(new PerpetualAccessEndIssueHandler());
-            eholdingMetaDataHandlers.add(new PerpetualAccessEndVolumeHandler());
-            eholdingMetaDataHandlers.add(new PersistentLinkHandler());
-            eholdingMetaDataHandlers.add(new PlatformHandler());
-            eholdingMetaDataHandlers.add(new ProxiedHandler());
-            eholdingMetaDataHandlers.add(new PublisherHandler());
-            eholdingMetaDataHandlers.add(new org.kuali.ole.dsng.rest.handler.eholdings.StatisticalSearchCodeHandler());
-            eholdingMetaDataHandlers.add(new SubscriptionStatusHandler());
-            eholdingMetaDataHandlers.add(new UrlHandler());
-            eholdingMetaDataHandlers.add(new LinkTextHandler());
-            eholdingMetaDataHandlers.add(new ImprintHandler());
-            eholdingMetaDataHandlers.add(new NonPublicNoteHandler());
-            eholdingMetaDataHandlers.add(new PublicNoteHandler());
-            eholdingMetaDataHandlers.add(new NoOfSumultaneousUserHander());
-        }
-        return eholdingMetaDataHandlers;
-    }
-
-    public void setBibHandlers(List<Handler> bibHandlers) {
-        this.bibHandlers = bibHandlers;
-    }
-
-    public List<Handler> getHoldingHandlers() {
-        if (null == holdingHandlers) {
-            holdingHandlers = new ArrayList<Handler>();
-            holdingHandlers.add(new CreateHoldingsHanlder());
-            holdingHandlers.add(new UpdateHoldingsHandler());
-        }
-        return holdingHandlers;
-    }
-
-    public List<Handler> getEHoldingHandlers() {
-        if (null == eholdingHandlers) {
-            eholdingHandlers = new ArrayList<Handler>();
-            eholdingHandlers.add(new CreateEHoldingsHandler());
-            eholdingHandlers.add(new UpdateEholdingsHandler());
-        }
-        return eholdingHandlers;
-    }
-
-    public void setHoldingHandlers(List<Handler> holdingHandlers) {
-        this.holdingHandlers = holdingHandlers;
-    }
-
-    public List<Handler> getItemHandlers() {
-        if (null == itemHandlers) {
-            itemHandlers = new ArrayList<Handler>();
-            itemHandlers.add(new CreateItemHandler());
-            itemHandlers.add(new UpdateItemHandler());
-        }
-        return itemHandlers;
-    }
-
-    public void setItemHandlers(List<Handler> itemHandlers) {
-        this.itemHandlers = itemHandlers;
-    }
-
     public String processBibAndHoldingsAndItems(String jsonBody) {
         String response = "{}";
-        OleNGBibImportResponse oleNGBibImportResponse = new OleNGBibImportResponse();
+
         List<BibResponse> bibResponses = new ArrayList<BibResponse>();
+
         Map<String, SolrInputDocument> solrInputDocumentMap = new HashedMap();
+
+        OleNGBibImportResponse oleNGBibImportResponse = new OleNGBibImportResponse();
+
         try {
             JSONArray requestJsonArray = new JSONArray(jsonBody);
             Exchange exchange = new Exchange();
+
             for (int index = 0; index < requestJsonArray.length(); index++) {
                 JSONObject bibJSONDataObject = requestJsonArray.getJSONObject(index);
 
                 BibResponse bibResponse = new BibResponse();
-                /*String valueOf001 = bibJSONDataObject.getString(OleNGConstants.TAG_001);
-                bibResponse.setValueOf001(valueOf001);*/
 
                 String ops = bibJSONDataObject.getString(OleNGConstants.OPS);
-
                 List<String> operationsList = getListFromJSONArray(ops);
 
                 BibRecord bibRecord = prepareBib(bibJSONDataObject);
 
-                if(StringUtils.isNotBlank(bibRecord.getBibId())) {
-                    bibResponse.setOperation(OleNGConstants.UPDATED);
-                } else {
-                    bibResponse.setOperation(OleNGConstants.CREATED);
-                }
+                updateOperationForBibResponse(bibResponse, bibRecord);
 
                 exchange.add(OleNGConstants.BIB, bibRecord);
 
                 List<HoldingsRecordAndDataMapping> holdingsForUpdateOrCreate = new ArrayList<HoldingsRecordAndDataMapping>();
-
 
                 preparePHoldingsRecord(bibRecord, bibJSONDataObject, exchange);
 
@@ -251,6 +86,8 @@ public class OleDsNgOverlayProcessor extends OleDsHelperUtil implements Docstore
                 List<HoldingsRecordAndDataMapping> updateHoldingsRecordAndDataMappings = (List<HoldingsRecordAndDataMapping>) exchange.get(OleNGConstants.HOLDINGS_FOR_UPDATE);
                 List<HoldingsRecordAndDataMapping> createEHoldingsRecordAndDataMappings = (List<HoldingsRecordAndDataMapping>) exchange.get(OleNGConstants.EHOLDINGS_FOR_CREATE);
                 List<HoldingsRecordAndDataMapping> updateEHoldingsRecordAndDataMappings = (List<HoldingsRecordAndDataMapping>) exchange.get(OleNGConstants.EHOLDINGS_FOR_UPDATE);
+
+
                 if (CollectionUtils.isNotEmpty(createHoldingsRecordAndDataMappings)) {
                     holdingsForUpdateOrCreate.addAll(createHoldingsRecordAndDataMappings);
                     oleNGBibImportResponse.setUnmatchedHoldingsCount(oleNGBibImportResponse.getUnmatchedHoldingsCount() + createHoldingsRecordAndDataMappings.size());
@@ -259,10 +96,10 @@ public class OleDsNgOverlayProcessor extends OleDsHelperUtil implements Docstore
                     holdingsForUpdateOrCreate.addAll(updateHoldingsRecordAndDataMappings);
                     oleNGBibImportResponse.setMatchedHoldingsCount(oleNGBibImportResponse.getMatchedHoldingsCount() + updateHoldingsRecordAndDataMappings.size());
                 }
-                if(CollectionUtils.isNotEmpty(createEHoldingsRecordAndDataMappings)) {
+                if (CollectionUtils.isNotEmpty(createEHoldingsRecordAndDataMappings)) {
                     oleNGBibImportResponse.setUnmatchedEHoldingsCount(oleNGBibImportResponse.getUnmatchedEHoldingsCount() + createEHoldingsRecordAndDataMappings.size());
                 }
-                if(CollectionUtils.isNotEmpty(updateEHoldingsRecordAndDataMappings)) {
+                if (CollectionUtils.isNotEmpty(updateEHoldingsRecordAndDataMappings)) {
                     oleNGBibImportResponse.setMatchedEHoldingsCount(oleNGBibImportResponse.getMatchedEHoldingsCount() + updateEHoldingsRecordAndDataMappings.size());
                 }
                 if (operationsList.contains("132") || operationsList.contains("231")) {
@@ -272,10 +109,10 @@ public class OleDsNgOverlayProcessor extends OleDsHelperUtil implements Docstore
                 List<ItemRecordAndDataMapping> createItemRecordAndDataMappings = (List<ItemRecordAndDataMapping>) exchange.get(OleNGConstants.ITEMS_FOR_CREATE);
                 List<ItemRecordAndDataMapping> updateItemRecordAndDataMappings = (List<ItemRecordAndDataMapping>) exchange.get(OleNGConstants.ITEMS_FOR_UPDATE);
 
-                if(CollectionUtils.isNotEmpty(createItemRecordAndDataMappings)) {
+                if (CollectionUtils.isNotEmpty(createItemRecordAndDataMappings)) {
                     oleNGBibImportResponse.setUnmatchedItemsCount(oleNGBibImportResponse.getUnmatchedItemsCount() + createItemRecordAndDataMappings.size());
                 }
-                if(CollectionUtils.isNotEmpty(updateItemRecordAndDataMappings)) {
+                if (CollectionUtils.isNotEmpty(updateItemRecordAndDataMappings)) {
                     oleNGBibImportResponse.setMatchedItemsCount(oleNGBibImportResponse.getMatchedItemsCount() + updateItemRecordAndDataMappings.size());
                 }
 
@@ -292,12 +129,12 @@ public class OleDsNgOverlayProcessor extends OleDsHelperUtil implements Docstore
                 buildBibResponses(bibResponse, exchange);
 
                 Integer multipleMatchedHoldings = (Integer) exchange.get("multipleMatchedHoldings");
-                if(null != multipleMatchedHoldings) {
+                if (null != multipleMatchedHoldings) {
                     oleNGBibImportResponse.setMultipleMatchedHoldingsCount(oleNGBibImportResponse.getMultipleMatchedHoldingsCount() + multipleMatchedHoldings);
                 }
 
                 Integer multipleMatchedEHoldings = (Integer) exchange.get("multipleMatchedEHoldings");
-                if(null != multipleMatchedEHoldings) {
+                if (null != multipleMatchedEHoldings) {
                     oleNGBibImportResponse.setMultipleMatchedEHoldingsCount(oleNGBibImportResponse.getMultipleMatchedEHoldingsCount() + multipleMatchedEHoldings);
                 }
 
@@ -323,17 +160,25 @@ public class OleDsNgOverlayProcessor extends OleDsHelperUtil implements Docstore
         return response;
     }
 
+    private void updateOperationForBibResponse(BibResponse bibResponse, BibRecord bibRecord) {
+        if (StringUtils.isNotBlank(bibRecord.getBibId())) {
+            bibResponse.setOperation(OleNGConstants.UPDATED);
+        } else {
+            bibResponse.setOperation(OleNGConstants.CREATED);
+        }
+    }
+
     public void buildBibResponses(BibResponse bibResponse, Exchange exchange) {
-        buildHoldingResponses(bibResponse,exchange);
-        buildEHoldingResponses(bibResponse,exchange);
+        buildHoldingResponses(bibResponse, exchange);
+        buildEHoldingResponses(bibResponse, exchange);
     }
 
     public void buildEHoldingResponses(BibResponse bibResponse, Exchange exchange) {
         List<HoldingsRecordAndDataMapping> createEHoldingsRecordAndDataMappings = (List<HoldingsRecordAndDataMapping>) exchange.get(OleNGConstants.EHOLDINGS_FOR_CREATE);
         List<HoldingsRecordAndDataMapping> updateEHoldingsRecordAndDataMappings = (List<HoldingsRecordAndDataMapping>) exchange.get(OleNGConstants.EHOLDINGS_FOR_UPDATE);
         List<HoldingsResponse> holdingsResponses = new ArrayList<HoldingsResponse>();
-        if(CollectionUtils.isNotEmpty(createEHoldingsRecordAndDataMappings)) {
-            for(HoldingsRecordAndDataMapping holdingsRecordAndDataMapping : createEHoldingsRecordAndDataMappings) {
+        if (CollectionUtils.isNotEmpty(createEHoldingsRecordAndDataMappings)) {
+            for (HoldingsRecordAndDataMapping holdingsRecordAndDataMapping : createEHoldingsRecordAndDataMappings) {
                 HoldingsResponse holdingsResponse = new HoldingsResponse();
                 HoldingsRecord holdingsRecord = holdingsRecordAndDataMapping.getHoldingsRecord();
                 holdingsResponse.setHoldingsId(holdingsRecord.getHoldingsId());
@@ -342,8 +187,8 @@ public class OleDsNgOverlayProcessor extends OleDsHelperUtil implements Docstore
                 holdingsResponses.add(holdingsResponse);
             }
         }
-        if(CollectionUtils.isNotEmpty(updateEHoldingsRecordAndDataMappings)) {
-            for(HoldingsRecordAndDataMapping holdingsRecordAndDataMapping : updateEHoldingsRecordAndDataMappings) {
+        if (CollectionUtils.isNotEmpty(updateEHoldingsRecordAndDataMappings)) {
+            for (HoldingsRecordAndDataMapping holdingsRecordAndDataMapping : updateEHoldingsRecordAndDataMappings) {
                 HoldingsResponse holdingsResponse = new HoldingsResponse();
                 HoldingsRecord holdingsRecord = holdingsRecordAndDataMapping.getHoldingsRecord();
                 holdingsResponse.setHoldingsId(holdingsRecord.getHoldingsId());
@@ -352,8 +197,8 @@ public class OleDsNgOverlayProcessor extends OleDsHelperUtil implements Docstore
                 holdingsResponses.add(holdingsResponse);
             }
         }
-        if(CollectionUtils.isNotEmpty(holdingsResponses)) {
-            if(CollectionUtils.isNotEmpty(bibResponse.getHoldingsResponses())) {
+        if (CollectionUtils.isNotEmpty(holdingsResponses)) {
+            if (CollectionUtils.isNotEmpty(bibResponse.getHoldingsResponses())) {
                 bibResponse.getHoldingsResponses().addAll(holdingsResponses);
             } else {
                 bibResponse.setHoldingsResponses(holdingsResponses);
@@ -365,11 +210,11 @@ public class OleDsNgOverlayProcessor extends OleDsHelperUtil implements Docstore
         List<HoldingsRecordAndDataMapping> createHoldingsRecordAndDataMappings = (List<HoldingsRecordAndDataMapping>) exchange.get(OleNGConstants.HOLDINGS_FOR_CREATE);
         List<HoldingsRecordAndDataMapping> updateHoldingsRecordAndDataMappings = (List<HoldingsRecordAndDataMapping>) exchange.get(OleNGConstants.HOLDINGS_FOR_UPDATE);
         List<HoldingsResponse> holdingsResponses = new ArrayList<HoldingsResponse>();
-        if(CollectionUtils.isNotEmpty(createHoldingsRecordAndDataMappings)) {
-            for(HoldingsRecordAndDataMapping holdingsRecordAndDataMapping : createHoldingsRecordAndDataMappings) {
+        if (CollectionUtils.isNotEmpty(createHoldingsRecordAndDataMappings)) {
+            for (HoldingsRecordAndDataMapping holdingsRecordAndDataMapping : createHoldingsRecordAndDataMappings) {
                 HoldingsResponse holdingsResponse = new HoldingsResponse();
                 HoldingsRecord holdingsRecord = holdingsRecordAndDataMapping.getHoldingsRecord();
-                if(null != holdingsRecord) {
+                if (null != holdingsRecord) {
                     holdingsResponse.setHoldingsId(holdingsRecord.getHoldingsId());
                     holdingsResponse.setOperation(OleNGConstants.CREATED);
                     holdingsResponse.setHoldingsType(PHoldings.PRINT);
@@ -379,11 +224,11 @@ public class OleDsNgOverlayProcessor extends OleDsHelperUtil implements Docstore
             }
         }
 
-        if(CollectionUtils.isNotEmpty(updateHoldingsRecordAndDataMappings)) {
-            for(HoldingsRecordAndDataMapping holdingsRecordAndDataMapping : updateHoldingsRecordAndDataMappings) {
+        if (CollectionUtils.isNotEmpty(updateHoldingsRecordAndDataMappings)) {
+            for (HoldingsRecordAndDataMapping holdingsRecordAndDataMapping : updateHoldingsRecordAndDataMappings) {
                 HoldingsResponse holdingsResponse = new HoldingsResponse();
                 HoldingsRecord holdingsRecord = holdingsRecordAndDataMapping.getHoldingsRecord();
-                if(null != holdingsRecord) {
+                if (null != holdingsRecord) {
                     holdingsResponse.setHoldingsId(holdingsRecord.getHoldingsId());
                     holdingsResponse.setOperation(OleNGConstants.UPDATED);
                     holdingsResponse.setHoldingsType(PHoldings.PRINT);
@@ -392,8 +237,8 @@ public class OleDsNgOverlayProcessor extends OleDsHelperUtil implements Docstore
                 }
             }
         }
-        if(CollectionUtils.isNotEmpty(holdingsResponses)) {
-            if(CollectionUtils.isNotEmpty(bibResponse.getHoldingsResponses())) {
+        if (CollectionUtils.isNotEmpty(holdingsResponses)) {
+            if (CollectionUtils.isNotEmpty(bibResponse.getHoldingsResponses())) {
                 bibResponse.getHoldingsResponses().addAll(holdingsResponses);
             } else {
                 bibResponse.setHoldingsResponses(holdingsResponses);
@@ -405,10 +250,10 @@ public class OleDsNgOverlayProcessor extends OleDsHelperUtil implements Docstore
         List<ItemRecordAndDataMapping> createItemRecordAndDataMappings = (List<ItemRecordAndDataMapping>) exchange.get(OleNGConstants.ITEMS_FOR_CREATE);
         List<ItemRecordAndDataMapping> updateItemRecordAndDataMappings = (List<ItemRecordAndDataMapping>) exchange.get(OleNGConstants.ITEMS_FOR_UPDATE);
         List<ItemResponse> itemResponses = new ArrayList<ItemResponse>();
-        if(CollectionUtils.isNotEmpty(createItemRecordAndDataMappings)) {
-            for(ItemRecordAndDataMapping itemRecordAndDataMapping : createItemRecordAndDataMappings) {
+        if (CollectionUtils.isNotEmpty(createItemRecordAndDataMappings)) {
+            for (ItemRecordAndDataMapping itemRecordAndDataMapping : createItemRecordAndDataMappings) {
                 ItemRecord itemRecord = itemRecordAndDataMapping.getItemRecord();
-                if(holdingsRecord.getHoldingsId().equals(itemRecord.getHoldingsId())) {
+                if (holdingsRecord.getHoldingsId().equals(itemRecord.getHoldingsId())) {
                     ItemResponse itemResponse = new ItemResponse();
                     itemResponse.setItemId(itemRecord.getItemId());
                     itemResponse.setOperation(OleNGConstants.CREATED);
@@ -416,11 +261,11 @@ public class OleDsNgOverlayProcessor extends OleDsHelperUtil implements Docstore
                 }
             }
         }
-        if(!(OleNGConstants.CREATED.equalsIgnoreCase(holdingsResponse.getOperation()))) {
-            if(CollectionUtils.isNotEmpty(updateItemRecordAndDataMappings)) {
-                for(ItemRecordAndDataMapping itemRecordAndDataMapping : updateItemRecordAndDataMappings) {
+        if (!(OleNGConstants.CREATED.equalsIgnoreCase(holdingsResponse.getOperation()))) {
+            if (CollectionUtils.isNotEmpty(updateItemRecordAndDataMappings)) {
+                for (ItemRecordAndDataMapping itemRecordAndDataMapping : updateItemRecordAndDataMappings) {
                     ItemRecord itemRecord = itemRecordAndDataMapping.getItemRecord();
-                    if(holdingsRecord.getHoldingsId().equals(itemRecord.getHoldingsId())) {
+                    if (holdingsRecord.getHoldingsId().equals(itemRecord.getHoldingsId())) {
                         ItemResponse itemResponse = new ItemResponse();
                         itemResponse.setItemId(itemRecord.getItemId());
                         itemResponse.setOperation(OleNGConstants.UPDATED);
@@ -429,8 +274,8 @@ public class OleDsNgOverlayProcessor extends OleDsHelperUtil implements Docstore
                 }
             }
         }
-        if(CollectionUtils.isNotEmpty(itemResponses)) {
-            if(CollectionUtils.isNotEmpty(holdingsResponse.getItemResponses())) {
+        if (CollectionUtils.isNotEmpty(itemResponses)) {
+            if (CollectionUtils.isNotEmpty(holdingsResponse.getItemResponses())) {
                 holdingsResponse.getItemResponses().addAll(itemResponses);
             } else {
                 holdingsResponse.setItemResponses(itemResponses);
@@ -558,7 +403,7 @@ public class OleDsNgOverlayProcessor extends OleDsHelperUtil implements Docstore
         Map<String, Integer> numOccurances = new TreeMap<String, Integer>(String.CASE_INSENSITIVE_ORDER);
         try {
             itemJSON = bibJSON.getJSONObject(OleNGConstants.ITEM);
-            numOccurances = getNumOccurances(bibJSON);
+            numOccurances = getNumOccurrences(bibJSON);
         } catch (JSONException e) {
             e.printStackTrace();
         }
@@ -733,14 +578,25 @@ public class OleDsNgOverlayProcessor extends OleDsHelperUtil implements Docstore
         return itemLinkField;
     }
 
+    /**
+     *
+     * @param bibRecord
+     * @param bibJSON
+     * @param exchange
+     *
+     * This method prepares a list of holdings to be either created or updated.
+     */
     private void preparePHoldingsRecord(BibRecord bibRecord, JSONObject bibJSON, Exchange exchange) {
         List<HoldingsRecordAndDataMapping> updatePHoldingsRecordAndDataMappings = new ArrayList<HoldingsRecordAndDataMapping>();
+
         List<HoldingsRecordAndDataMapping> createPHoldingsRecordAndDataMappings = new ArrayList<HoldingsRecordAndDataMapping>();
+
         JSONObject holdingsJSON = null;
+
         Map<String, Integer> numOccurances = new TreeMap<String, Integer>(String.CASE_INSENSITIVE_ORDER);
         try {
             holdingsJSON = bibJSON.getJSONObject(OleNGConstants.HOLDINGS);
-            numOccurances = getNumOccurances(bibJSON);
+            numOccurances = getNumOccurrences(bibJSON);
 
         } catch (JSONException e) {
             e.printStackTrace();
@@ -779,12 +635,15 @@ public class OleDsNgOverlayProcessor extends OleDsHelperUtil implements Docstore
 
     public void prepareEHoldingsRecord(BibRecord bibRecord, JSONObject bibJSON, Exchange exchange) {
         List<HoldingsRecordAndDataMapping> updateEHoldingsRecordAndDataMappings = new ArrayList<HoldingsRecordAndDataMapping>();
+
         List<HoldingsRecordAndDataMapping> createEHoldingsRecordAndDataMappings = new ArrayList<HoldingsRecordAndDataMapping>();
+
         JSONObject eholdingsJSON = null;
+
         Map<String, Integer> numOccurances = new TreeMap<String, Integer>(String.CASE_INSENSITIVE_ORDER);
         try {
             eholdingsJSON = bibJSON.getJSONObject(OleNGConstants.EHOLDINGS);
-            numOccurances = getNumOccurances(bibJSON);
+            numOccurances = getNumOccurrences(bibJSON);
 
         } catch (JSONException e) {
             e.printStackTrace();
@@ -835,6 +694,7 @@ public class OleDsNgOverlayProcessor extends OleDsHelperUtil implements Docstore
 
     public List<HoldingsRecordAndDataMapping> determineHoldingsAndDataMappingsByMatchPoints(BibRecord bibRecord, Exchange exchange, JSONObject holdingsJSON, String docType) {
         List<HoldingsRecordAndDataMapping> holdingsRecordAndDataMappings = new ArrayList<HoldingsRecordAndDataMapping>();
+
         List<HoldingsRecord> holdingsForBib = bibRecord.getHoldingsRecords();
 
         JSONArray clonedDataMapping = getClonedDataMappingsJSONArray(holdingsJSON);
@@ -845,62 +705,45 @@ public class OleDsNgOverlayProcessor extends OleDsHelperUtil implements Docstore
             HoldingsRecord holdingsRecord = iterator.next();
             if (holdingsRecord.getHoldingsType().equals(docType)) {
 
-                JSONObject matchPoints = null;
+                JSONObject matchPoints = getJsonObject(holdingsJSON, OleNGConstants.MATCH_POINT);
 
+                HashMap map = getHashMapFromObjectMapper(matchPoints);
 
-                try {
-                    matchPoints = holdingsJSON.getJSONObject(OleNGConstants.MATCH_POINT);
+                for (Iterator matchPointIterator = map.keySet().iterator(); matchPointIterator.hasNext(); ) {
+                    String key = (String) matchPointIterator.next();
 
-                    HashMap map = getObjectMapper().readValue(matchPoints.toString(), new TypeReference<Map<String, String>>() {
-                    });
+                    for (Iterator<HoldingsHandler> holdingsMetaDataHandlerIterator = getHoldingMetaDataHandlersIterator(docType); holdingsMetaDataHandlerIterator.hasNext(); ) {
+                        HoldingsHandler holdingsMetadataHandler = holdingsMetaDataHandlerIterator.next();
 
-                    for (Iterator matchPointIterator = map.keySet().iterator(); matchPointIterator.hasNext(); ) {
-                        String key = (String) matchPointIterator.next();
-                        Iterator<HoldingsHandler> holdingsMetaDataHandlerIterator = getHoldingMetaDataHandlers(docType);
+                        if (holdingsMetadataHandler.isInterested(key)) {
+                            exchange.add(OleNGConstants.HOLDINGS_RECORD, holdingsRecord);
 
+                            holdingsMetadataHandler.process(matchPoints, exchange);
 
-                        for (Iterator<HoldingsHandler> holdingsRecordIterator = holdingsMetaDataHandlerIterator; holdingsRecordIterator.hasNext(); ) {
-                            HoldingsHandler holdingsHandler = holdingsRecordIterator.next();
+                            Object match = exchange.get(OleNGConstants.MATCHED_HOLDINGS);
 
-                            if (holdingsHandler.isInterested(key)) {
-                                exchange.add(OleNGConstants.HOLDINGS_RECORD, holdingsRecord);
+                            if (null != match && match.equals(Boolean.TRUE)) {
+                                HoldingsRecordAndDataMapping holdingsRecordAndDataMapping = null;
+                                holdingsRecordAndDataMapping = new HoldingsRecordAndDataMapping();
+                                holdingsRecordAndDataMapping.setHoldingsRecord(holdingsRecord);
+                                String matchedValue = (String) exchange.get(OleNGConstants.MATCHED_VALUE);
+                                JSONObject filterdDataMapping = findDataMappingByValue(clonedDataMapping, key, matchedValue);
+                                holdingsRecordAndDataMapping.setDataMapping(filterdDataMapping);
 
-                                holdingsHandler.process(matchPoints, exchange);
-
-                                Object match = exchange.get(OleNGConstants.MATCHED_HOLDINGS);
-
-                                if (null != match && match.equals(Boolean.TRUE)) {
-                                    HoldingsRecordAndDataMapping holdingsRecordAndDataMapping = null;
-                                    holdingsRecordAndDataMapping = new HoldingsRecordAndDataMapping();
-                                    holdingsRecordAndDataMapping.setHoldingsRecord(holdingsRecord);
-                                    String matchedValue = (String) exchange.get(OleNGConstants.MATCHED_VALUE);
-                                    JSONObject filterdDataMapping = findDataMappingByValue(clonedDataMapping, key, matchedValue);
-                                    holdingsRecordAndDataMapping.setDataMapping(filterdDataMapping);
-
-                                    if (!matchedHoldingsByMatchPoint.containsKey(key)) {
-                                        ArrayList matchedHoldingsForKey = new ArrayList();
-                                        matchedHoldingsForKey.add(holdingsRecordAndDataMapping);
-                                        matchedHoldingsByMatchPoint.put(key, matchedHoldingsForKey);
-                                    } else {
-                                        List<HoldingsRecordAndDataMapping> holdingsRecordAndDataMappings1 = matchedHoldingsByMatchPoint.get(key);
-                                        holdingsRecordAndDataMappings1.add(holdingsRecordAndDataMapping);
-                                    }
-
-                                    exchange.remove(OleNGConstants.MATCHED_HOLDINGS);
-                                    exchange.remove(OleNGConstants.MATCHED_VALUE);
+                                if (!matchedHoldingsByMatchPoint.containsKey(key)) {
+                                    ArrayList matchedHoldingsForKey = new ArrayList();
+                                    matchedHoldingsForKey.add(holdingsRecordAndDataMapping);
+                                    matchedHoldingsByMatchPoint.put(key, matchedHoldingsForKey);
+                                } else {
+                                    List<HoldingsRecordAndDataMapping> holdingsRecordAndDataMappings1 = matchedHoldingsByMatchPoint.get(key);
+                                    holdingsRecordAndDataMappings1.add(holdingsRecordAndDataMapping);
                                 }
+
+                                exchange.remove(OleNGConstants.MATCHED_HOLDINGS);
+                                exchange.remove(OleNGConstants.MATCHED_VALUE);
                             }
                         }
                     }
-
-                } catch (JSONException e) {
-                    e.printStackTrace();
-                } catch (JsonParseException e) {
-                    e.printStackTrace();
-                } catch (JsonMappingException e) {
-                    e.printStackTrace();
-                } catch (IOException e) {
-                    e.printStackTrace();
                 }
             }
 
@@ -909,20 +752,20 @@ public class OleDsNgOverlayProcessor extends OleDsHelperUtil implements Docstore
 
         for (Iterator<String> holdingsRecordIterator = matchedHoldingsByMatchPoint.keySet().iterator(); holdingsRecordIterator.hasNext(); ) {
             String key = holdingsRecordIterator.next();
-            if(matchedHoldingsByMatchPoint.get(key).size() == 1){
+            if (matchedHoldingsByMatchPoint.get(key).size() == 1) {
                 holdingsRecordAndDataMappings.add(matchedHoldingsByMatchPoint.get(key).get(0));
             } else {
-                if(docType.equalsIgnoreCase(PHoldings.PRINT)) {
+                if (docType.equalsIgnoreCase(PHoldings.PRINT)) {
                     Integer multipleMatchedHoldings = (Integer) exchange.get("multipleMatchedHoldings");
-                    if(null != multipleMatchedHoldings) {
+                    if (null != multipleMatchedHoldings) {
                         multipleMatchedHoldings = multipleMatchedHoldings + matchedHoldingsByMatchPoint.size();
                     } else {
                         multipleMatchedHoldings = matchedHoldingsByMatchPoint.size();
                     }
                     exchange.add("multipleMatchedHoldings", multipleMatchedHoldings);
-                } else if(docType.equalsIgnoreCase(EHoldings.ELECTRONIC)) {
+                } else if (docType.equalsIgnoreCase(EHoldings.ELECTRONIC)) {
                     Integer multipleMatchedEHoldings = (Integer) exchange.get("multipleMatchedEHoldings");
-                    if(null != multipleMatchedEHoldings) {
+                    if (null != multipleMatchedEHoldings) {
                         multipleMatchedEHoldings = multipleMatchedEHoldings + matchedHoldingsByMatchPoint.size();
                     } else {
                         multipleMatchedEHoldings = matchedHoldingsByMatchPoint.size();
@@ -936,7 +779,7 @@ public class OleDsNgOverlayProcessor extends OleDsHelperUtil implements Docstore
         if (CollectionUtils.isNotEmpty(unAssignedDataMappings)) {
             for (Iterator<JSONObject> iterator = unAssignedDataMappings.iterator(); iterator.hasNext(); ) {
                 JSONObject dataMapping = iterator.next();
-                HoldingsRecordAndDataMapping newEHoldings = getNewHoldings(bibRecord, EHoldings.ELECTRONIC);
+                HoldingsRecordAndDataMapping newEHoldings = getNewHoldings(bibRecord, docType);
                 newEHoldings.setDataMapping(dataMapping);
                 holdingsRecordAndDataMappings.add(newEHoldings);
             }
@@ -945,7 +788,28 @@ public class OleDsNgOverlayProcessor extends OleDsHelperUtil implements Docstore
         return holdingsRecordAndDataMappings;
     }
 
-    private Iterator<HoldingsHandler> getHoldingMetaDataHandlers(String docType) {
+    private HashMap getHashMapFromObjectMapper(JSONObject jsonObject) {
+        if (null != jsonObject) {
+            try {
+                return getObjectMapper().readValue(jsonObject.toString(), new TypeReference<Map<String, String>>() {
+                });
+            } catch (IOException e) {
+                e.printStackTrace();
+            }
+        }
+        return new HashMap();
+    }
+
+    private JSONObject getJsonObject(JSONObject jsonObject, String key) {
+        try {
+            return jsonObject.getJSONObject(key);
+        } catch (JSONException e) {
+            e.printStackTrace();
+        }
+        return null;
+    }
+
+    private Iterator<HoldingsHandler> getHoldingMetaDataHandlersIterator(String docType) {
         if (docType.equalsIgnoreCase(PHoldings.PRINT)) {
             return getHoldingMetaDataHandlers().iterator();
         } else if (docType.equalsIgnoreCase(EHoldings.ELECTRONIC)) {
@@ -1001,7 +865,13 @@ public class OleDsNgOverlayProcessor extends OleDsHelperUtil implements Docstore
         return unAssignedDataMappings;
     }
 
-    private Map<String, Integer> getNumOccurances(JSONObject bibJSON) throws JSONException {
+    /**
+     * @param bibJSON
+     * @return
+     * @throws JSONException The method calculates the num occurrences of a tag in the given bib record. i.e. how many times
+     *                       has 856 tag appears. Primarily for setting up how many holdings/e-holdings/items to create.
+     */
+    private Map<String, Integer> getNumOccurrences(JSONObject bibJSON) throws JSONException {
         Map numOccurances = new TreeMap<String, Integer>(String.CASE_INSENSITIVE_ORDER);
         numOccurances.put(OleNGConstants.HOLDINGS, 1);
         numOccurances.put(OleNGConstants.EHOLDINGS, 1);
@@ -1114,20 +984,24 @@ public class OleDsNgOverlayProcessor extends OleDsHelperUtil implements Docstore
         return returnValue;
     }
 
-    private JSONObject findDataMappingByValue(JSONArray dataMappings,
-                                              String type, String value) throws JSONException {
+    private JSONObject findDataMappingByValue(JSONArray dataMappings, String type, String value) {
         if (null != dataMappings) {
             for (int index = 0; index < dataMappings.length(); index++) {
-                JSONObject dataMapping = dataMappings.getJSONObject(index);
-                if (dataMapping.has(type)) {
-                    JSONArray dataMappingValues = dataMapping.getJSONArray(type);
-                    for (int innerIndex = 0; innerIndex < dataMappingValues.length(); innerIndex++) {
-                        String dataMappingValue = dataMappingValues.getString(innerIndex);
-                        if (dataMappingValue.equalsIgnoreCase(value)) {
-                            dataMapping.put("assigned", Boolean.TRUE);
-                            return dataMapping;
+                JSONObject dataMapping = null;
+                try {
+                    dataMapping = dataMappings.getJSONObject(index);
+                    if (dataMapping.has(type)) {
+                        JSONArray dataMappingValues = dataMapping.getJSONArray(type);
+                        for (int innerIndex = 0; innerIndex < dataMappingValues.length(); innerIndex++) {
+                            String dataMappingValue = dataMappingValues.getString(innerIndex);
+                            if (dataMappingValue.equalsIgnoreCase(value)) {
+                                dataMapping.put("assigned", Boolean.TRUE);
+                                return dataMapping;
+                            }
                         }
                     }
+                } catch (JSONException e) {
+                    e.printStackTrace();
                 }
             }
         }
